inventory_app/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
distlib==0.3.1
Django==3.1.1
django-crispy-forms==1.9.2
django-softdelete==0.9.1
docker==3.4.1
docker-compose==1.21.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
idna==2.6
importlib-metadata==1.7.0
jsonschema==2.6.0
keyring==17.1.1
keyrings.alt==3.1.1
pipenv==2020.8.13
psycopg2-binary==2.8.5
pycrypto==2.6.1
PyGObject==3.30.4
pytz==2020.1
pyxdg==0.25
PyYAML==3.13
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
sqlparse==0.3.1
texttable==1.6.0
urllib3==1.24.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
websocket-client==0.53.0
zipp==3.1.0
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
distlib==0.3.1
Django==3.1.1
django-admin-tools==0.9.0
django-crispy-forms==1.9.2
django-fluent-dashboard==1.0.1
django-softdelete==0.9.1
docker==3.4.1
docker-compose==1.21.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.6
importlib-metadata==1.7.0
jsonschema==2.6.0
keyring==17.1.1
keyrings.alt==3.1.1
pipenv==2020.8.13
psycopg2-binary==2.8.5
pycrypto==2.6.1
PyGObject==3.30.4
pytz==2020.1
pyxdg==0.25
PyYAML==3.13
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
sqlparse==0.3.1
texttable==1.6.0
urllib3==1.24.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
websocket-client==0.53.0
zipp==3.1.0
